# RoamOpenAPI/Sources/RoamOpenAPI/openapi.yaml
{
  "openapi": "3.0.3",
  "info": {
    "title": "Roam API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/ideas": {
      "get": {
        "summary": "Listideas",
        "operationId": "listIdeas_api_ideas_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/IdeaRead"
                  },
                  "type": "array",
                  "title": "Response Listideas Api Ideas Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Createidea",
        "operationId": "createIdea_api_ideas_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IdeaCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IdeaRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/ideas/{ideaId}": {
      "get": {
        "summary": "Getidea",
        "operationId": "getIdea_api_ideas__ideaId__get",
        "parameters": [
          {
            "name": "ideaId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Ideaid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IdeaRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "summary": "Updateidea",
        "operationId": "updateIdea_api_ideas__ideaId__patch",
        "parameters": [
          {
            "name": "ideaId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Ideaid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IdeaUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IdeaRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Deleteidea",
        "operationId": "deleteIdea_api_ideas__ideaId__delete",
        "parameters": [
          {
            "name": "ideaId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Ideaid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/ideas/{ideaId}/interpret": {
      "post": {
        "summary": "Interpretidea",
        "operationId": "interpretIdea_api_ideas__ideaId__interpret_post",
        "parameters": [
          {
            "name": "ideaId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Ideaid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IdeaRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/ideas/{ideaId}/place-suggestions/{suggestionId}": {
      "patch": {
        "summary": "Selectplacesuggestion",
        "operationId": "selectPlaceSuggestion_api_ideas__ideaId__place_suggestions__suggestionId__patch",
        "parameters": [
          {
            "name": "ideaId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Ideaid"
            }
          },
          {
            "name": "suggestionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Suggestionid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IdeaRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/ideas/{ideaId}/suggest-slots": {
      "post": {
        "summary": "Suggestslotsforidea",
        "description": "Generate suggested time slots for an idea (on demand, not stored).",
        "operationId": "suggestSlotsForIdea_api_ideas__ideaId__suggest_slots_post",
        "parameters": [
          {
            "name": "ideaId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Ideaid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Suggestslotsforidea Api Ideas  Ideaid  Suggest Slots Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/ideas/{ideaId}/plan": {
      "post": {
        "summary": "Promoteideatoplan",
        "operationId": "promoteIdeaToPlan_api_ideas__ideaId__plan_post",
        "parameters": [
          {
            "name": "ideaId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Ideaid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "IdeaCreate": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          },
          "notes": {
            "type": "string",
            "title": "Notes",
            "default": ""
          },
          "sourceUrl": {
            "type": "string",
            "title": "Sourceurl",
            "default": ""
          },
          "displayName": {
            "type": "string",
            "nullable": true,
            "title": "Displayname"
          }
        },
        "type": "object",
        "required": [
          "title"
        ],
        "title": "IdeaCreate"
      },
      "IdeaRead": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "userId": {
            "type": "string",
            "format": "uuid",
            "title": "Userid"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "notes": {
            "type": "string",
            "title": "Notes"
          },
          "sourceUrl": {
            "type": "string",
            "title": "Sourceurl"
          },
          "displayName": {
            "type": "string",
            "nullable": true,
            "title": "Displayname"
          },
          "placeId": {
            "type": "string",
            "format": "uuid",
            "nullable": true,
            "title": "Placeid"
          },
          "savedReelId": {
            "type": "string",
            "format": "uuid",
            "nullable": true,
            "title": "Savedreelid"
          },
          "status": {
            "$ref": "#/components/schemas/IdeaStatus"
          },
          "planCount": {
            "type": "integer",
            "title": "Plancount",
            "default": 0
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "title": "Createdat"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "title": "Updatedat"
          },
          "pipelineResult": {
            "$ref": "#/components/schemas/PipelineResultRead",
            "nullable": true
          },
          "placeName": {
            "type": "string",
            "nullable": true,
            "title": "Placename"
          },
          "placeLatitude": {
            "type": "number",
            "nullable": true,
            "title": "Placelatitude"
          },
          "placeLongitude": {
            "type": "number",
            "nullable": true,
            "title": "Placelongitude"
          }
        },
        "type": "object",
        "required": [
          "id",
          "userId",
          "title",
          "notes",
          "sourceUrl",
          "status",
          "createdAt",
          "updatedAt"
        ],
        "title": "IdeaRead"
      },
      "IdeaStatus": {
        "type": "string",
        "enum": [
          "captured",
          "suggesting",
          "ready",
          "planned"
        ],
        "title": "IdeaStatus"
      },
      "IdeaUpdate": {
        "properties": {
          "title": {
            "type": "string",
            "nullable": true,
            "title": "Title"
          },
          "notes": {
            "type": "string",
            "nullable": true,
            "title": "Notes"
          },
          "sourceUrl": {
            "type": "string",
            "nullable": true,
            "title": "Sourceurl"
          },
          "displayName": {
            "type": "string",
            "nullable": true,
            "title": "Displayname"
          },
          "status": {
            "$ref": "#/components/schemas/IdeaStatus",
            "nullable": true
          },
          "placeId": {
            "type": "string",
            "format": "uuid",
            "nullable": true,
            "title": "Placeid"
          }
        },
        "type": "object",
        "title": "IdeaUpdate"
      },
...
      "PipelineResultRead": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "ideaId": {
            "type": "string",
            "format": "uuid",
            "nullable": true,
            "title": "Ideaid"
          },
          "jobId": {
            "type": "string",
            "format": "uuid",
            "nullable": true,
            "title": "Jobid"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "refinedTitle": {
            "type": "string",
            "nullable": true,
            "title": "Refinedtitle"
          },
          "category": {
            "type": "string",
            "nullable": true,
            "title": "Category"
          },
          "estimatedMinutes": {
            "type": "integer",
            "nullable": true,
            "title": "Estimatedminutes"
          },
          "modelName": {
            "type": "string",
            "nullable": true,
            "title": "Modelname"
          },
          "promptVersion": {
            "type": "string",
            "nullable": true,
            "title": "Promptversion"
          },
          "rawOutput": {
            "additionalProperties": true,
            "type": "object",
            "nullable": true,
            "title": "Rawoutput"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "title": "Createdat"
          },
          "placeSuggestions": {
            "items": {
              "$ref": "#/components/schemas/PlaceSuggestionRead"
            },
            "type": "array",
            "title": "Placesuggestions",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "id",
          "source",
          "createdAt"
        ],
        "title": "PipelineResultRead"
      },
...
      "PlanRead": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "creatorId": {
            "type": "string",
            "format": "uuid",
            "title": "Creatorid"
          },
          "placeId": {
            "type": "string",
            "format": "uuid",
            "nullable": true,
            "title": "Placeid"
          },
          "ideaId": {
            "type": "string",
            "format": "uuid",
            "nullable": true,
            "title": "Ideaid"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "scheduledAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true,
            "title": "Scheduledat"
          },
          "status": {
            "$ref": "#/components/schemas/PlanStatus"
          },
          "shareCode": {
            "type": "string",
            "title": "Sharecode"
          },
          "rawPrompt": {
            "type": "string",
            "nullable": true,
            "title": "Rawprompt"
          },
          "estimatedMinutes": {
            "type": "integer",
            "title": "Estimatedminutes"
          },
          "calendarEventId": {
            "type": "string",
            "nullable": true,
            "title": "Calendareventid"
          },
          "partySize": {
            "type": "integer",
            "title": "Partysize"
          },
          "rating": {
            "type": "integer",
            "nullable": true,
            "title": "Rating"
          },
          "taskNotes": {
            "type": "string",
            "title": "Tasknotes"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "title": "Createdat"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "title": "Updatedat"
          },
          "members": {
            "items": {
              "$ref": "#/components/schemas/PlanMemberRead"
            },
            "type": "array",
            "title": "Members",
            "default": []
          },
          "placeName": {
            "type": "string",
            "nullable": true,
            "title": "Placename"
          }
        },
        "type": "object",
        "required": [
          "id",
          "creatorId",
          "title",
          "status",
          "shareCode",
          "estimatedMinutes",
          "partySize",
          "taskNotes",
          "createdAt",
          "updatedAt"
        ],
        "title": "PlanRead"
      },
      "PlanStatus": {
        "type": "string",
        "enum": [
          "draft",
          "confirmed",
          "completed",
          "cancelled"
        ],
        "title": "PlanStatus"
      },
...
    }
  }
}